--- code/firebase/database/build.gradle.kts
@@ -1,25 +1,0 @@
-plugins {
-    id (Dependencies.Plugin.Library)
-    kotlin(Dependencies.Plugin.Android)
-    kotlin(Dependencies.Plugin.Kapt)
-    id(Dependencies.Plugin.DaggerHilt)
-}
-
-android {
-    compileSdk = Dependencies.Android.CompileSdkVersion
-
-    defaultConfig {
-        minSdk = Dependencies.Android.MinSdkVersion
-    }
-}
-
-dependencies {
-    implementation(Dependencies.Libs.Dagger.Android)
-
-    implementation(Dependencies.Libs.Firebase.Database)
-    implementation(Dependencies.Libs.Kotlin.Coroutines)
-
-    implementation(Dependencies.Libs.Hilt.DaggerHiltAndroid)
-    kapt(Dependencies.Libs.Hilt.DaggerHiltAndroidCompiler)
-    kapt(Dependencies.Libs.Hilt.hiltCompiler)
-}
