--- code/firebase/database/src/main/java/com/nicholasrutherford/chal/firebase/database/ChalFirebaseDatabase.kt
@@ -1,9 +1,0 @@
-package com.nicholasrutherford.chal.firebase.database
-
-import com.google.firebase.database.DatabaseReference
-import com.google.firebase.database.FirebaseDatabase
-
-interface ChalFirebaseDatabase {
-    val database: FirebaseDatabase
-    fun userDatabaseReference(uid: String): DatabaseReference
-}
