+++ code/firebase/realtime-database/fetch/src/main/java/com/nicholasrutherford/chal/firebase/realtime/database/fetch/FetchFirebaseDatabase.kt
@@ -1,0 +1,10 @@
+package com.nicholasrutherford.chal.firebase.realtime.database.fetch
+
+import com.google.firebase.database.DatabaseReference
+import com.google.firebase.database.FirebaseDatabase
+
+interface FetchFirebaseDatabase {
+    val database: FirebaseDatabase
+
+    fun fetchAllUsersDatabaseReference(uid: String): DatabaseReference
+}
