+++ code/firebase/realtime-database/fetch/src/main/java/com/nicholasrutherford/chal/firebase/realtime/database/fetch/FetchFirebaseDatabaseImpl.kt
@@ -1,0 +1,15 @@
+package com.nicholasrutherford.chal.firebase.realtime.database.fetch
+
+import com.google.firebase.database.DatabaseReference
+import com.google.firebase.database.FirebaseDatabase
+import com.nicholasrutherford.chal.helper.constants.userDatabaseReference
+import javax.inject.Inject
+
+class FetchFirebaseDatabaseImpl @Inject constructor(): FetchFirebaseDatabase {
+
+    override val database = FirebaseDatabase.getInstance()
+
+    override fun fetchAllUsersDatabaseReference(uid: String): DatabaseReference {
+        return database.getReference(userDatabaseReference(uid))
+    }
+}
