+++ code/settings.gradle.kts
@@ -21,7 +21,7 @@ include(":shared-preference:create")
 include(":shared-preference:fetch")
 include(":shared-preference:remove")
 include(":firebase:storage")
-include(":firebase:database")
 include(":data:account-info")
 include(":main:news-feed")
 include(":helper:constants")
+include(":firebase:realtime-database:fetch")
